python_app/requirements.txt
fastapi
uvicorn
# OpenTelemetry
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-distro
opentelemetry-instrumentation-system-metrics
psutil